requirements.txt
canonicalwebteam.discourse==4.0.9
canonicalwebteam.flask-base==1.0.5
canonicalwebteam.search==1.2.0
canonicalwebteam.templatefinder==1.0.0
canonicalwebteam.versioned-static==1.0.2
vcrpy-unittest==0.1.7
